+++ src/app/modules/profile/services/profile.service.ts
@@ -1,26 +1,30 @@
-import { HttpClient, HttpHeaders, HttpResponse } from '@angular/common/http';
+import {
+    HttpClient,
+    HttpHeaders,
+    HttpParams,
+    HttpResponse,
+} from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { AppConfigService } from '@app/core/services/app-config/app-config.service';
+import { statusReturn } from '@app/shared/interfaces/deployment.interface';
+import {
+    RequestLoginResponse,
+    CompleteLoginResponse,
+    StorageCredential,
+} from '@app/shared/interfaces/profile.interface';
+import { environment } from '@environments/environment';
 import { Observable } from 'rxjs';
 
-export interface RequestLoginResponse {
-    poll: {
-        token: string;
-        endpoint: string;
-    };
-    login: string;
-}
-
-export interface CompleteLoginResponse {
-    server: string;
-    loginName: string;
-    appPassword: string;
-}
+const { base, endpoints } = environment.api;
 
 @Injectable({
     providedIn: 'root',
 })
 export class ProfileService {
-    constructor(private http: HttpClient) {}
+    constructor(
+        private http: HttpClient,
+        private appConfigService: AppConfigService
+    ) {}
 
     initLogin(domain: string): Observable<RequestLoginResponse> {
         const url = 'https://' + domain + '/index.php/login/v2';
@@ -28,7 +32,7 @@ export class ProfileService {
         return this.http.post<RequestLoginResponse>(url, body);
     }
 
-    getCredentials(
+    getNewCredentials(
         loginResponse: RequestLoginResponse
     ): Observable<HttpResponse<CompleteLoginResponse>> {
         const url = loginResponse.poll.endpoint;
@@ -40,6 +44,42 @@ export class ProfileService {
         return this.http.post<CompleteLoginResponse>(url, body, {
             headers,
             observe: 'response',
+        });
+    }
+
+    getExistingCredentials(): Observable<StorageCredential[]> {
+        const url = `${base}${endpoints.secrets}`;
+        const params = new HttpParams()
+            .set('vo', this.appConfigService.voName)
+            .set('subpath', '/services/storage');
+
+        return this.http.get<Array<StorageCredential>>(url, {
+            params: params,
+        });
+    }
+
+    addCredential(
+        credential: StorageCredential,
+        serviceName: string
+    ): Observable<statusReturn> {
+        const url = `${base}${endpoints.secrets}`;
+        const params = new HttpParams()
+            .set('vo', this.appConfigService.voName)
+            .set('secret_path', '/services/storage/' + serviceName);
+
+        return this.http.post<statusReturn>(url, credential, {
+            params: params,
+        });
+    }
+
+    deleteCredential(serviceName: string): Observable<statusReturn> {
+        const url = `${base}${endpoints.secrets}`;
+        const params = new HttpParams()
+            .set('vo', this.appConfigService.voName)
+            .set('secret_path', '/services/storage/' + serviceName);
+
+        return this.http.delete<statusReturn>(url, {
+            params: params,
         });
     }
 }